% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_1_1_calculateModels.R
\name{calculateModels}
\alias{calculateModels}
\title{calculateModels}
\usage{
calculateModels(
  processedData,
  numberOfClusters = NA,
  minWordsPerCluster = 5,
  maxWordsPerCluster = 10,
  p = 0.05,
  dendrogram = TRUE,
  dendroLabels = "truncated",
  generateWordlist = FALSE,
  saveToWd = TRUE,
  ordinationFunction = FALSE,
  longMessages = FALSE
)
}
\arguments{
\item{processedData}{result of \code{\link[scicloud]{processMetaDataMatrix}}}

\item{numberOfClusters}{integer or NA; an integer forces the number of clusters,
NA results in an automatic iteration to determine the optimal amount.}

\item{minWordsPerCluster}{minimum number of words to be plotted per cluster.}

\item{maxWordsPerCluster}{maximum number of words to be plotted per cluster.
this is only happening if the indicator values are higher than the maximum
value of the weakest cluster arm.}

\item{p}{p-Value to determine the significance of individual words. Only
significant words will be plotted.}

\item{dendrogram}{shows a dendrogram of the calculated cluster if set to \code{TRUE}.}

\item{dendroLabels}{allows "truncated" or "break" to either truncate the
labels of the dendrogram leaves or put a line break. Line breaks are not
recommend for a large number of PDFs.}

\item{generateWordlist}{if set to \code{TRUE}, it generates a wordlist in
your working directory. The list contains all significant words that the
indicator species analysis deemed significant to describe your paper
clusters. To work with the new wordlist, read it in using
\code{keepWordsFile} as an argument to \code{\link[scicloud]{ordinationCluster()}}.}

\item{saveToWd}{a logical parameter whether or not to save the output of the
function to the working directory. This is especially useful for later
analysis steps and can be read in by using \code{\link[base]{saveRDS}}.}

\item{ordinationFunction}{placeholder}

\item{longMessages}{placeholder}
}
\value{
list of two entries. First entry contains all the information to
    generate the ordination/cluster plot. Entry two is an updated version
    of the metaMatrix, which now included the assigned cluster number per
    paper to allow additional statistical analyses.
}
\description{
The third function to the word analysis with scicloud. It takes the
    output of \code{\link[scicloud]{processMetaDataMatrix}} and calculates
    ordination and cluster models. Each paper is assigned to one cluster
    while each word receives an indicator value with \code{\link[labdsv]{indval}}
    for each cluster, showing how representative word is for a cluster. The top
    representative words will be used in the further process with \code{
    \link[scicloud]{createOrdinationPlot}}.
}
\examples{
\dontrun{ placeholder }
}
\seealso{
\code{\link{processMetaDataMatrix}} for the preceding step,
    \code{\link{createOrdinationPlot}} for the graphics,
    \code{\link{mostImportantPaperPerCluster}} and
    \code{\link{inspectScicloud}} for a summary of the analysis

Other scicloud functions: 
\code{\link{createOrdinationPlot}()},
\code{\link{createTextMatrixFromPDF}()},
\code{\link{getScopusMetaData}()},
\code{\link{inspectScicloud}()},
\code{\link{mostImportantPaperPerCluster}()},
\code{\link{ordinationCluster}()},
\code{\link{processMetaDataMatrix}()},
\code{\link{searchScopus}()}
}
\author{
Matthias Nachtmann, \email{matthias.nachtmann@stud.leuphana.de},
    Lisa Gotzian, \email{lisa.gotzian@stud.leuphana.de}
}
\concept{scicloud functions}
