% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_2_calculateNetwork.R
\name{calculateNetwork}
\alias{calculateNetwork}
\title{calculateNetwork}
\usage{
calculateNetwork(
  processedMetaMatrix,
  sortby = c("Eigenvector", "Degree", "Closeness", "Betweenness"),
  keep = 0.33,
  saveToWd = TRUE,
  ordinationFunction = FALSE,
  longMessages = FALSE
)
}
\arguments{
\item{processedMetaMatrix}{requires the output of
\code{\link[scicloud]{processMetaDataMatrix}}.}

\item{sortby}{allows the following possible inputs: "Eigenvector", "Degree",
"Closeness, "Betweenness". The centrality measure to sort the words by,
default is Eigenvector.}

\item{keep}{The argument keep keeps by default 0.33 of all the words, sorted
by the argument given by \code{sortby}. Can be adjusted. This facilitates
computations for later use.}

\item{saveToWd}{placeholder}

\item{ordinationFunction}{placeholder}

\item{longMessages}{placeholder}
}
\value{
output:
    $LocalMeasures will return the local measurements for both papers and
    words
    $ReducedLocalMeasures will return 1/3 of the words (!) with their
    centrality measures & clustering according to three different clustering
    methods, arranged by eigenvector centrality (can be changed)
    $ReducedIncidenceMatrix will return 1/3 of the words arranged by
    eigenvector centrality, to be further processed e.g. in Gephi or with other
    clustering functions
    $GlobalMeasures will return the global measurements
}
\description{
This function is an alternative to
    \code{\link[scicloud]{calculateModels}} and clusters based on network theory.
}
\seealso{
\code{\link{processMetaDataMatrix}} for the preceding step,
    \code{\link{inspectScicloud}} for a summary of the analysis
}
\author{
Lisa Gotzian, \email{lisa.gotzian@stud.leuphana.de}, Julius Rathgens,
    \email{julius.rathgens@leuphana.de}
}
