% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_1_2_createOrdinationPlot.R
\name{createOrdinationPlot}
\alias{createOrdinationPlot}
\title{createOrdinationPlot}
\usage{
createOrdinationPlot(
  modeledData,
  exactPosition = FALSE,
  ordinationFunction = FALSE
)
}
\arguments{
\item{modeledData}{result of \code{\link[scicloud]{calculateModels}}}

\item{exactPosition}{\code{TRUE} or \code{FALSE}, the plot function tries
to avoid overlapping labels for the sake of visual simplicity over perfect
precision. When set to \code{TRUE}, the words position will be marked by
a dot and the label will be connected with a line to it.}

\item{ordinationFunction}{for internal use.}
}
\value{
a graphic based on the calculated model and some additional barplot
    to deepen the understanding of the dataset.
}
\description{
The fourth function to the word analysis with scicloud. It creates
    five different plots that set the results in context to the years
    published and more. To show meaningful graphics, the use of
    \code{\link[scicloud]{getScopusMetaData}} is recommended.
}
\examples{
\dontrun{
placeholder}
}
\seealso{
\code{\link[scicloud]{calculateModels}} for the preceding step,
    \code{\link[scicloud]{mostImportantPaperPerCluster}} and
    \code{\link[scicloud]{inspectScicloud}} for a summary of the analysis

Other scicloud functions: 
\code{\link{calculateModels}()},
\code{\link{createTextMatrixFromPDF}()},
\code{\link{getScopusMetaData}()},
\code{\link{inspectScicloud}()},
\code{\link{mostImportantPaperPerCluster}()},
\code{\link{ordinationCluster}()},
\code{\link{processMetaDataMatrix}()},
\code{\link{searchScopus}()}
}
\author{
Matthias Nachtmann, \email{matthias.nachtmann@stud.leuphana.de},
    Lisa Gotzian, \email{lisa.gotzian@stud.leuphana.de}
}
\concept{scicloud functions}
