% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createScipusList.R
\name{createScipusList}
\alias{createScipusList}
\title{createScipusList}
\usage{
createScipusList(
  directory = file.path(".", "PDFs"),
  myAPIKey = NA,
  language = "SMART",
  stemWords = TRUE,
  saveToWd = FALSE,
  long_msg = FALSE,
  ignoreWords = c(),
  keepWordsFile = NA
)
}
\description{
createScipusList
}
