% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_createTextMatrixFromPDF.R
\name{createTextMatrixFromPDF}
\alias{createTextMatrixFromPDF}
\title{createTextMatrixFromPDF}
\usage{
createTextMatrixFromPDF(directory = file.path(".", "PDFs"), saveToWd = TRUE)
}
\arguments{
\item{directory}{per default, the PDFs are expected to be in a folder named
"PDFs", can be changed ad. lib.}

\item{saveToWd}{a logical parameter whether or not to save the output of the
function to the working directory. This is especially useful for later
analysis steps and can be read in by using \code{\link[base]{readRDS}}.}
}
\value{
A data frame containing the file name and full text of the pdf, as
    well as scraped DOI numbers from the text and some placeholder columns for later.
}
\description{
First function of the word analysis with scicloud. It takes all
    scientific papers as PDF files from a "PDFs" folder you'll have to
    create. It then creates a DocumentTerm matrix
    similar to \code{\link[scicloud]{searchScopus}} for further processing.
}
\examples{
\dontrun{
metaMatrix <- createTextMatrixFromPDF(saveToWd = TRUE)

myAPIKey <- "YOUR API KEY"
# go to https://dev.elsevier.com/user/login to get yours

scicloudAnalysis <- ordinationCluster(metaMatrix, myAPIKey = myAPIKey, stemWords = FALSE,
    longMessages = TRUE, saveToWd = TRUE, method = "hclust")
scicloudSpecs <- inspectScicloud(modeledData = scicloudAnalysis)
}
}
\seealso{
\code{\link[scicloud]{ordinationCluster}} for the next step in scicloud,
    a wrapper of all steps or \code{\link[scicloud]{processMetaDataMatrix}} for the
    next step if you intend to run it step by step,
    \code{\link[scicloud]{getScopusMetaData}} to fill in paper metadata (needed for
    future plots) from Scopus

Other scicloud functions: 
\code{\link{calculateModels}()},
\code{\link{createOrdinationPlot}()},
\code{\link{delete_RDS}()},
\code{\link{getScopusMetaData}()},
\code{\link{inspectScicloud}()},
\code{\link{mostImportantPaperPerCluster}()},
\code{\link{ordinationCluster}()},
\code{\link{processMetaDataMatrix}()},
\code{\link{searchScopus}()}
}
\author{
Matthias Nachtmann, \email{matthias.nachtmann@stud.leuphana.de},
    Lisa Gotzian, \email{lisa.gotzian@stud.leuphana.de}
}
\concept{scicloud functions}
