% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{delete_RDS}
\alias{delete_RDS}
\title{delete_RDS}
\usage{
delete_RDS()
}
\value{

}
\description{
delete RDS files from current working directory. Users can
     choose 3 delete options from the menu
     Option 1: Delete All RDS files in cwd 
     Option 2: Keep the latest RDS files of each data type 
     Option 3: Keep only the selected data type of RDS files, delete all others
}
\seealso{


Other scicloud functions: 
\code{\link{calculateModels}()},
\code{\link{createOrdinationPlot}()},
\code{\link{createTextMatrixFromPDF}()},
\code{\link{getScopusMetaData}()},
\code{\link{inspectScicloud}()},
\code{\link{mostImportantPaperPerCluster}()},
\code{\link{ordinationCluster}()},
\code{\link{processMetaDataMatrix}()},
\code{\link{searchScopus}()}
}
\author{
Jia Yan Ng, \email{jia.y.ng@stud.leuphana.de}
}
\concept{scicloud functions}
