% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_till_4_1_3_ordinationCluster.R
\name{ordinationCluster}
\alias{ordinationCluster}
\title{ordinationCluster}
\usage{
ordinationCluster(
  metaMatrix,
  language = "SMART",
  numberOfClusters = NA,
  minWordsPerCluster = 5,
  maxWordsPerCluster = 10,
  stemWords = TRUE,
  ignoreWords = c(),
  myAPIKey = NA,
  p = 0.05,
  exactPosition = FALSE,
  dendrogram = TRUE,
  dendroLabels = "truncated",
  generateWordlist = FALSE,
  keepWordsFile,
  saveToWd = TRUE,
  method = "hclust"
)
}
\arguments{
\item{metaMatrix}{placeholder}

\item{language}{placeholder}

\item{numberOfClusters}{placeholder}

\item{minWordsPerCluster}{placeholder}

\item{maxWordsPerCluster}{placeholder}

\item{stemWords}{placeholder}

\item{ignoreWords}{placeholder}

\item{myAPIKey}{your private Elsevier API key for communicating with the
API. You can request one at \url{https://dev.elsevier.com/index.jsp}.}

\item{p}{placeholder}

\item{exactPosition}{placeholder}

\item{dendrogram}{placeholder}

\item{dendroLabels}{placeholder}

\item{generateWordlist}{logical, if set to \code{TRUE}, it generates a wordlist in
your working directory. The list contains all significant words that the
indicator species analysis deemed significant to describe your paper
clusters. You can now add a 0/1 behind each word or delete rows you
don't consider important to the analysis. To re-run the analysis with
the new wordlist, read it in using \code{keepWordsFile} as an argument
to \code{\link{ordinationCluster}}.}

\item{keepWordsFile}{path to a .csv-file that specifies which words to keep
during the analysis. Accepts 0/1 behind each word or takes the words
as they are and disregards all other words of the analysis.}

\item{saveToWd}{a logical parameter whether or not to save the output of the
function to the working directory. This is especially useful for later
analysis steps. The file can be read in by using \code{\link[base]{readRDS}}.}

\item{method}{takes "network", "hclust" or "both" as a method}
}
\value{
placeholder
}
\description{
The essential wrapper function that runs all steps of the word
    analysis.
}
\examples{
\dontrun{

### The normal workflow of scicloud
myAPIKey <- "YOUR_API_KEY"
metaMatrix <- createTextMatrixFromPDF()


# run the analysis
scicloudAnalysis <- ordinationCluster(metaMatrix,
                           myAPIKey = myAPIKey)

# inspect the analysis
scicloudSpecs <- inspectScicloud(scicloudAnalysis)


### Workflow for generating a wordlist to revise the words
scicloudAnalysis <- ordinationCluster(metaMatrix, 
                                      generateWordlist = TRUE,
                                      stemWords = TRUE, numberOfClusters = 4)
scicloudAnalysis <- ordinationCluster(metaMatrix,
                                      keepWordsFile = "YourFile.csv",
                                      stemWords = TRUE, numberOfClusters = 4)
# inspect the analysis
scicloudSpecs <- inspectScicloud(scicloudAnalysis)
}
}
\seealso{
Other scicloud functions: 
\code{\link{calculateModels}()},
\code{\link{createOrdinationPlot}()},
\code{\link{createTextMatrixFromPDF}()},
\code{\link{delete_RDS}()},
\code{\link{getScopusMetaData}()},
\code{\link{inspectScicloud}()},
\code{\link{mostImportantPaperPerCluster}()},
\code{\link{processMetaDataMatrix}()},
\code{\link{searchScopus}()}
}
\author{
Matthias Nachtmann, \email{matthias.nachtmann@stud.leuphana.de},
    Lisa Gotzian, \email{lisa.gotzian@stud.leuphana.de}
}
\concept{scicloud functions}
