% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_processMetaDataMatrix.R
\name{processMetaDataMatrix}
\alias{processMetaDataMatrix}
\title{processMetaDataMatrix}
\usage{
processMetaDataMatrix(
  metaMatrix,
  control = list(),
  ignoreWords = c(),
  keepWordsFile
)
}
\arguments{
\item{metaMatrix}{metaMatrix created through \code{\link[ginko]{getScopusMetaData}}
or \code{\link[ginko]{createTextMatrixFromPDF}}. Equations, symbols, all words
in parentheses and all references are removed.}

\item{control}{a list of parameters used to determine pre-processing methods.
Error will be thrown if language & stemWords are not defined.
language: this defines the stopwords to be filtered. the default is
"english". Look at \code{\link[tm]{stopwords}} for more information.
stemWords: can be \code{TRUE} of \code{FALSE}. Transforms every word
to its stem, so variants of the same words are treated equally. Look
at \code{\link[tm]{stemDocument}} for more information.
saveToWd: a logical parameter whether or not to save the output of the function to the working directory
ordinationFunction placeholder
e.g. control = list(language = "SMART", stemWords = FALSE, saveToWd = TRUE, ordinationFunction = FALSE)}

\item{ignoreWords}{a vector of words to be ignored.}

\item{keepWordsFile}{path to a .csv-file that specifies which words to keep
during the analysis. If provided, all other words will be disregarded.}
}
\value{
returns a list object with \code{[["Tf_idf"]]} as the tf-idf document
    term matrix, \code{[["MetaMatrix"]]} as passed to the function and
\code{[["wordList"]]} is the list of words found in the papers.
}
\description{
The second function to the word analysis with ginko. This function
    accepts a \code{metaMatrix} and constructs a tf-idf matrix.
}
\seealso{
\code{\link{createTextMatrixFromPDF}} for the preceding step,
    \code{\link{calculateModels}} for the proceeding step

Other ginko functions: 
\code{\link{calculateModels}()},
\code{\link{createOrdinationPlot}()},
\code{\link{createTextMatrixFromPDF}()},
\code{\link{inspectGinko}()},
\code{\link{mostImportantPaperPerCluster}()},
\code{\link{ordinationCluster}()}
}
\author{
Jia Yan Ng, \email{jia.y.ng@stud.leuphana.de}
}
\concept{ginko functions}
