% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rename_PDFs}
\alias{rename_PDFs}
\title{rename_PDFs}
\usage{
rename_PDFs()
}
\value{
placeholder
}
\description{
rename PDF files in order to avoid errors due to non-standard characters
}
\seealso{
placeholder

Other scicloud functions: 
\code{\link{calculateModels}()},
\code{\link{createOrdinationPlot}()},
\code{\link{createTextMatrixFromPDF}()},
\code{\link{delete_RDS}()},
\code{\link{getScopusMetaData}()},
\code{\link{inspectScicloud}()},
\code{\link{mostImportantPaperPerCluster}()},
\code{\link{ordinationCluster}()},
\code{\link{processMetaDataMatrix}()},
\code{\link{searchScopus}()}
}
\author{
Johann Julius Beeck, \email{johann.j.beeck@stud.leuphana.de}
}
\concept{scicloud functions}
