% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runAnalysis.R
\name{runAnalysis}
\alias{runAnalysis}
\title{runAnalysis}
\usage{
runAnalysis(
  scipusList,
  numberOfClusters = NA,
  minWordsPerCluster = 5,
  maxWordsPerCluster = 10,
  p = 0.05,
  exactPosition = FALSE,
  dendrogram = TRUE,
  dendroLabels = c("truncated", "break"),
  saveToWd = TRUE,
  method = c("hclust", "network", "both"),
  generateWordlist = FALSE,
  long_msg = FALSE
)
}
\description{
runAnalysis
}
