% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchScopus.R
\name{searchScopus}
\alias{searchScopus}
\title{searchScopus}
\usage{
searchScopus(
  searchString,
  myAPIKey,
  maxResults = 10,
  countIncrement = 200,
  saveToWd = TRUE
)
}
\arguments{
\item{searchString}{The search string you want to ask the server. See the
Scopus API (\url{https://dev.elsevier.com/index.jsp/}) for details.}

\item{myAPIKey}{your private Elsevier API key for communicating with the
API. You can request one at \url{https://dev.elsevier.com/index.jsp}.}

\item{maxResults}{The maximum amount of accepted search results. Usually,
Scopus does not provide more than 5000 results.}

\item{countIncrement}{The number of results per GET request. A private user
can't exceed 25 per request. If you are inside a subscribed IP range,
you can use the maximum of 200 per request. Note, that the weekly quota
for requests is 20,000.}

\item{saveToWd}{a logical parameter whether or not to save the output of the
function to the working directory. This is especially useful for later
analysis steps. The file can be read in by using \code{\link[base]{readRDS}}.}
}
\value{
A data frame containing the DOI numbers and Scopus-IDs of the search
    results, as well as paper metadata like author and publishing year.
}
\description{
This function accepts a search string in URL format and an
    Elsevier API Key and returns the DOI numbers of all search results as
    a vector for further processing.
}
\examples{
\dontrun{
DOInumbers <- searchScopus("TITLE-ABS-KEY(sustainability) AND PUBYEAR > 2009",
  "1234567890ABCDEF",
  maxResults = 160, countIncrement = 20
)
DOInumbers
}

}
\seealso{
Other scicloud functions: 
\code{\link{createScicloudList}()},
\code{\link{deleteRDS}()},
\code{\link{inspectScicloud}()},
\code{\link{runAnalysis}()}
}
\author{
Creator of the scicloud workflow: Henrik von Wehrden,
  \email{henrik.von_wehrden@leuphana.de} \cr \cr
  Code by: Matthias Nachtmann,
  \email{matthias.nachtmann@stud.leuphana.de},
  Lisa Gotzian, \email{lisa.gotzian@stud.leuphana.de},
  Prabesh Dhakal, \email{prabesh.dhakal@stud.leuphana.de} \cr \cr
  First version of scicloud: Matthias Nachtmann,
  \email{matthias.nachtmann@stud.leuphana.de}
}
\concept{scicloud functions}
