% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_2_calculateNetwork.R
\name{calculateNetwork}
\alias{calculateNetwork}
\title{calculateNetwork}
\usage{
calculateNetwork(
  processedMetaDataMatrix,
  sortby = c("Eigenvector", "Degree", "Closeness", "Betweenness"),
  keep = 0.33,
  saveToWd = TRUE,
  ordinationFunction = FALSE
)
}
\arguments{
\item{processedMetaDataMatrix}{result of \code{\link{processMetaDataMatrix}}}

\item{sortby}{the centrality measure to sort the words by,
default is Eigenvector. Allows the following possible inputs: "Eigenvector", "Degree",
"Closeness, "Betweenness".}

\item{keep}{numeric, keeps by default 0.33 of all the words, sorted
by the argument given by \code{sortby}. A smaller amount of words to
keep facilitates computations for later use.}

\item{saveToWd}{a logical parameter whether or not to save the output of the
function to the working directory. This is especially useful for later
analysis steps. The file can be read in by using \code{\link[base]{readRDS}}.}

\item{ordinationFunction}{internal variable}
}
\value{
output:
\itemize{
    \item \code{LocalMeasures}: local measures
    for both papers and words
    \item \code{ReducedLocalMeasures}: 1/3 of the words (!) with their
    centrality measures & clustering according to three different clustering
    methods, arranged by default by eigenvector centrality using \code{sortby}
    \item \code{ReducedIncidenceMatrix}: 1/3 of the words arranged by
    eigenvector centrality, to be further processed e.g. in Gephi or with other
    clustering functions
    \item \code{GlobalMeasures}: global measures of the network
    }
}
\description{
This function is an alternative to the fourth function
    \code{\link{calculateModels}} of the scicloud analysis. It uses
    a network clustering approach. When done, it returns a list of global
    and local measures and also generates a clustered matrix. This matrix
    can then be further processed in network programs like
    Gephi.
}
\examples{
\dontrun{

### The normal workflow of scicloud
myAPIKey <- "YOUR_API_KEY"
metaMatrix <- createTextMatrixFromPDF()


# run the analysis
scicloudAnalysis <- ordinationCluster(metaMatrix,
                           myAPIKey = myAPIKey,
                           method = "network")

# inspect the analysis
scicloudSpecs <- inspectScicloud(scicloudAnalysis)
}
}
\seealso{
\itemize{
    \item \code{\link{processMetaDataMatrix}} for the preceding step
    \item \code{\link{inspectScicloud}} for a summary of the analysis
    \item note that \code{\link{createOrdinationPlot}}, 
    \code{\link{mostImportantPaperPerCluster}} and \code{\link{inspectScicloud}}
    are part of the clustering with \code{\link{calculateModels}}
    and don't work with this function.
    }
}
\author{
Lisa Gotzian, \email{lisa.gotzian@stud.leuphana.de}, Julius Rathgens,
    \email{julius.rathgens@leuphana.de}
}
