% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_createTextMatrixFromPDF.R
\name{createTextMatrixFromPDF}
\alias{createTextMatrixFromPDF}
\title{createTextMatrixFromPDF}
\usage{
createTextMatrixFromPDF(directory = file.path(".", "PDFs"), saveToWd = TRUE)
}
\arguments{
\item{directory}{per default, the PDFs are expected to be in a folder named
"PDFs", can be changed ad. lib.}

\item{saveToWd}{a logical parameter whether or not to save the output of the
function to the working directory. This is especially useful for later
analysis steps. The file can be read in by using \code{\link[base]{readRDS}}.}
}
\value{
A data frame containing the file name and full text of the pdf, 
    the DOI numbers from the text and some empty metadata columns to be
    filled by \code{\link{getScopusMetaData}}.
    It is analogous to what \code{\link{searchScopus}}
    returns.
}
\description{
First function of the word analysis with scicloud. It takes all
    scientific papers as PDF files from the "PDFs" folder in your working
    directory. It then creates a DocumentTerm matrix of it.
}
\examples{
\dontrun{

### The normal workflow of scicloud
myAPIKey <- "YOUR_API_KEY"
metaMatrix <- createTextMatrixFromPDF()


# run the analysis, see ordinationCluster()
# for more arguments
scicloudAnalysis <- ordinationCluster(metaMatrix,
                           myAPIKey = myAPIKey)

# inspect the analysis
scicloudSpecs <- inspectScicloud(scicloudAnalysis)
}
}
\seealso{
\itemize{
    \item \code{\link{rename_PDFs}} in case of errors due to vulnerable file names
    \item \code{\link{ordinationCluster}} for the next step in scicloud
    \item or \code{\link{processMetaDataMatrix}} for the
    more granular next step if you intend to run it step by step
    \item \code{\link{getScopusMetaData}} to fill in paper metadata (needed for
    future plots) from Scopus
    }

Other scicloud functions: 
\code{\link{calculateModels}()},
\code{\link{createOrdinationPlot}()},
\code{\link{delete_RDS}()},
\code{\link{getScopusMetaData}()},
\code{\link{inspectScicloud}()},
\code{\link{mostImportantPaperPerCluster}()},
\code{\link{ordinationCluster}()},
\code{\link{processMetaDataMatrix}()},
\code{\link{searchScopus}()}
}
\author{
Matthias Nachtmann, \email{matthias.nachtmann@stud.leuphana.de},
    Lisa Gotzian, \email{lisa.gotzian@stud.leuphana.de}
}
\concept{scicloud functions}
