% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_getScopusMetaData.R
\name{getScopusMetaData}
\alias{getScopusMetaData}
\title{getScopusMetaData}
\usage{
getScopusMetaData(metaMatrix, myAPIKey = NA, long_msg = FALSE)
}
\arguments{
\item{metaMatrix}{a dataframe generated
either from PDFs with \code{\link{createTextMatrixFromPDF}} or by
searching Scopus with \code{\link{searchScopus}}}

\item{myAPIKey}{your private Elsevier API key for communicating with the
API. You can request one at \url{https://dev.elsevier.com/index.jsp}.}

\item{long_msg}{logical variable to whether print long message or not}
}
\value{
A dataframe with metadata of all articles provided to the function.
}
\description{
The second function to the word analysis with scicloud. This
    function accepts a dataframe of DOI numbers and/or Scopus-IDs.
    It downloads article metadata like title, author and year and returns
    the input dataframe filled with that metadata. Make sure to be connected
    to your institution's network (e.g. via VPN).
}
\examples{
\dontrun{

### The normal workflow of scicloud
myAPIKey <- "YOUR_API_KEY"
metaMatrix <- createTextMatrixFromPDF()


# instead of ordinationCluster(), we can also run this
# workflow step by step.

# 1) pull article metadata from Scopus
metaMatrix <- getScopusMetaData(metaMatrix, myAPIKey)

# 2) process the full texts
processedMetaDataMatrix <- processMetaDataMatrix(
          metaMatrix,
          list(language = "SMART",
          stemWords = TRUE,
          saveToWd = FALSE),
          ignoreWords = c("Abstract", "Bulletin", "Editor"))

# 3) run the cluster analysis to determine publication communities
scicloudAnalysis <- calculateModels(processedMetaDataMatrix)

# 4) visualize the results
createOrdinationPlot(scicloudAnalysis)

# 5) a list of the most important papers per cluster
mostImportantPaperPerCluster(scicloudAnalysis)

# 6) a summary of the analysis
scicloudSpecs <- inspectScicloud(scicloudAnalysis)
    }
}
\seealso{
\itemize{
    \item \code{\link{createTextMatrixFromPDF}} or \code{\link{searchScopus}}
    for the preceding steps
    \item \code{\link{processMetaDataMatrix}} for the proceeding step
    }

Other scicloud functions: 
\code{\link{calculateModels}()},
\code{\link{createOrdinationPlot}()},
\code{\link{createTextMatrixFromPDF}()},
\code{\link{createTfIdf}()},
\code{\link{delete_RDS}()},
\code{\link{inspectScicloud}()},
\code{\link{mostImportantPaperPerCluster}()},
\code{\link{ordinationCluster}()},
\code{\link{processMetaDataMatrix}()},
\code{\link{searchScopus}()}
}
\author{
Matthias Nachtmann, \email{matthias.nachtmann@stud.leuphana.de},
    Lisa Gotzian, \email{lisa.gotzian@stud.leuphana.de}, Prabesh Dhakal,
    \email{prabesh.dhakal@stud.leuphana.de}
}
\concept{scicloud functions}
