% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_1_4_inspectScicloud.R
\name{inspectScicloud}
\alias{inspectScicloud}
\title{inspectScicloud}
\usage{
inspectScicloud(modeledData = modeledData)
}
\arguments{
\item{modeledData}{placeholder}
}
\value{
placeholder
}
\description{
placeholder
}
\seealso{
\code{\link{calculateModels}} or \code{\link{calculateNetwork}}
    for the preceding step,
    \code{\link{createOrdinationPlot}} for the graphics,
    \code{\link{mostImportantPaperPerCluster}}

Other scicloud functions: 
\code{\link{calculateModels}()},
\code{\link{createOrdinationPlot}()},
\code{\link{createTextMatrixFromPDF}()},
\code{\link{getScopusMetaData}()},
\code{\link{mostImportantPaperPerCluster}()},
\code{\link{ordinationCluster}()},
\code{\link{processMetaDataMatrix}()},
\code{\link{searchScopus}()}
}
\author{
Lisa Gotzian, \email{lisa.gotzian@stud.leuphana.de}
}
\concept{scicloud functions}
