% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspectScicloud.R
\name{inspectScicloud}
\alias{inspectScicloud}
\title{Summarize the result of scicloud analysis}
\usage{
inspectScicloud(scicloudAnalysis)
}
\arguments{
\item{scicloudAnalysis}{}
}
\value{
A summary of the analysis is printed into the console. It gives
    insights into the total number of papers and words used.
    The returned object contains:
    \itemize{
    \item \code{paperCluster}: a paper-cluster table. Each paper
    belongs to one cluster. Use \code{View(scicloudSpecs$paperCluster)} to see
    which paper belongs to which cluster.
    \item \code{IndVal}: a words-cluster table. Words do not
    belong to one single cluster. An indicator species analysis shows how
    representative each word is for each cluster. Use \code{View(scicloudSpecs$IndVal)}
    to view the results of the indicator species analysis.
    \item \code{excludedPapers}: the excluded papers. Use 
    \code{View(scicloudSpecs$excludedPapers)} to see which papers have
    been excluded, possibly because of a PDF error.
    \item \code{representativePapers}: the most representative papers. Use 
    \code{View(scicloudSpecs$representativePapers)} to see which papers
    are the most representative ones, weighted with the indicator species
    values of the words in the paper.
    \item \code{metaMatrix}: matrix with metadata of the papers. Use
    \code{View(scicloudSpecs$metaMatrix)} to view the original
    metaMatrix (without full texts, so it's safe to open).
    }
}
\description{
The last function to be called to perform analysis in scicloud.
It takes the result of \code{scicloudAnalysis} and returns a summary of the 
cluster analysis.
}
\examples{
\dontrun{

### Workflow of performing analysis using scicloud
myAPIKey <- "YOUR_API_KEY"
# retrieving data from PDFs and Scorpus website using API  
scicloudList <- createScicloudList(myAPIKey = myAPIKey)

# Run the analysis with a specified no. of cluster
scicloudAnalysis <- runAnalysis(scipusList = scipusList, numberOfClusters = 4)
                           
# Generate a summary of the analysis 
scicloudSpecs <- inspectScicloud(scicloudAnalysis)
}
}
\seealso{
\itemize{
    \item \code{\link{calculateModels}} for the preceding step
    \item \code{\link{createOrdinationPlot}} for the graphics
    \item \code{\link{mostImportantPaperPerCluster}} for the most cited
    papers per cluster
    }

Other scicloud functions: 
\code{\link{createScicloudList}()},
\code{\link{deleteRDS}()},
\code{\link{runAnalysis}()}
}
\author{
Lisa Gotzian, \email{lisa.gotzian@stud.leuphana.de}
}
\concept{scicloud functions}
