% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_1_3_mostImportantPaperPerCluster.R
\name{mostImportantPaperPerCluster}
\alias{mostImportantPaperPerCluster}
\title{mostImportantPaperPerCluster}
\usage{
mostImportantPaperPerCluster(modeledData)
}
\arguments{
\item{modeledData}{result of \code{\link[scicloud]{calculateModels}}}
}
\value{
console output of the papers with the highest citation per year
    count per cluster.
}
\description{
placeholder
}
\examples{
\dontrun{
placeholder}
}
\seealso{
\code{\link[scicloud]{calculateModels}} for the preceding step
    \code{\link[scicloud]{createOrdinationPlot}} for the graphics,
    \code{\link[scicloud]{inspectScicloud}} for a summary of the analysis

Other scicloud functions: 
\code{\link{calculateModels}()},
\code{\link{createOrdinationPlot}()},
\code{\link{createTextMatrixFromPDF}()},
\code{\link{delete_RDS}()},
\code{\link{getScopusMetaData}()},
\code{\link{inspectScicloud}()},
\code{\link{ordinationCluster}()},
\code{\link{processMetaDataMatrix}()},
\code{\link{searchScopus}()}
}
\author{
Matthias Nachtmann, \email{matthias.nachtmann@stud.leuphana.de},
    Lisa Gotzian, \email{lisa.gotzian@stud.leuphana.de}
}
\concept{scicloud functions}
