% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_1_3_mostImportantPaperPerCluster.R
\name{mostImportantPaperPerCluster}
\alias{mostImportantPaperPerCluster}
\title{mostImportantPaperPerCluster}
\usage{
mostImportantPaperPerCluster(scicloudAnalysis)
}
\arguments{
\item{scicloudAnalysis}{result of \code{\link{calculateModels}}}
}
\value{
This function returns a console output of the papers with the
    highest citation count per year per cluster. The citations were fetched
    from the Scopus API by \code{\link{getScopusMetaData}}.
}
\description{
The sixth function to the word analysis with scicloud. It takes
    \code{scicloudAnalysis} and outputs a list of the most cited papers per
    cluster into the console.
}
\examples{
\dontrun{

### The normal workflow of scicloud
myAPIKey <- "YOUR_API_KEY"
metaMatrix <- createTextMatrixFromPDF()


# instead of ordinationCluster(), we can also run this
# workflow step by step.

# 1) pull article metadata from Scopus
metaMatrix <- getScopusMetaData(metaMatrix, myAPIKey)

# 2) process the full texts
processedMetaDataMatrix <- processMetaDataMatrix(
          metaMatrix,
          list(language = "SMART",
          stemWords = TRUE,
          saveToWd = FALSE),
          ignoreWords = c("Abstract", "Bulletin", "Editor"))
                                  
# 3) run the cluster analysis to determine publication communities
scicloudAnalysis <- calculateModels(processedMetaDataMatrix)

# 4) visualize the results
createOrdinationPlot(scicloudAnalysis)

# 5) a list of the most important papers per cluster
mostImportantPaperPerCluster(scicloudAnalysis)

# 6) a summary of the analysis
scicloudSpecs <- inspectScicloud(scicloudAnalysis)
    }
}
\seealso{
\itemize{
    \item \code{\link{calculateModels}} for the preceding step
    \item \code{\link{createOrdinationPlot}} for the graphics
    \item \code{\link{inspectScicloud}} for a summary of the analysis
    }

Other scicloud functions: 
\code{\link{calculateModels}()},
\code{\link{createOrdinationPlot}()},
\code{\link{createTextMatrixFromPDF}()},
\code{\link{delete_RDS}()},
\code{\link{getScopusMetaData}()},
\code{\link{inspectScicloud}()},
\code{\link{ordinationCluster}()},
\code{\link{processMetaDataMatrix}()},
\code{\link{rename_PDFs}()},
\code{\link{searchScopus}()}
}
\author{
Matthias Nachtmann, \email{matthias.nachtmann@stud.leuphana.de},
    Lisa Gotzian, \email{lisa.gotzian@stud.leuphana.de}
}
\concept{scicloud functions}
