% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_1_3_mostImportantPaperPerCluster.R
\name{mostImportantPaperPerCluster}
\alias{mostImportantPaperPerCluster}
\title{mostImportantPaperPerCluster}
\usage{
mostImportantPaperPerCluster(modeledData)
}
\arguments{
\item{modeledData}{result of \code{\link[ginko]{calculateModels}}}
}
\value{
console output of the papers with the hightest citation per year
    count per cluster.
}
\description{
placeholder
}
\examples{
\dontrun{
placeholder}
}
\seealso{
\code{\link{calculateModels}} for the preceding step
    \code{\link{createOrdinationPlot}} for the graphics,
    \code{\link{inspectGinko}} for a summary of the analysis

Other ginko functions: 
\code{\link{calculateModels}()},
\code{\link{createOrdinationPlot}()},
\code{\link{createTextMatrixFromPDF}()},
\code{\link{inspectGinko}()},
\code{\link{ordinationCluster}()},
\code{\link{processMetaDataMatrix}()}
}
\author{
Matthias Nachtmann, \email{matthias.nachtmann@stud.leuphana.de},
    Lisa Gotzian, \email{lisa.gotzian@stud.leuphana.de}
}
\concept{ginko functions}
