% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_till_4_1_3_ordinationCluster.R
\name{ordinationCluster}
\alias{ordinationCluster}
\title{ordinationCluster}
\usage{
ordinationCluster(
  metaMatrix,
  language = "SMART",
  numberOfClusters = NA,
  minWordsPerCluster = 5,
  maxWordsPerCluster = 10,
  stemWords = TRUE,
  ignoreWords = c(),
  myAPIKey = NA,
  p = 0.05,
  exactPosition = FALSE,
  dendrogram = TRUE,
  dendroLabels = "truncated",
  saveToWd = TRUE,
  method = "hclust",
  longMessages = FALSE
)
}
\arguments{
\item{metaMatrix}{placeholder}

\item{language}{placeholder}

\item{numberOfClusters}{placeholder}

\item{minWordsPerCluster}{placeholder}

\item{maxWordsPerCluster}{placeholder}

\item{stemWords}{placeholder}

\item{ignoreWords}{placeholder}

\item{myAPIKey}{placeholder}

\item{p}{placeholder}

\item{exactPosition}{placeholder}

\item{dendrogram}{placeholder}

\item{dendroLabels}{placeholder}

\item{saveToWd}{a logical parameter whether or not to save the output of
the function to the working directory. This is especially useful for
later analysis steps and can be read in by using \code{\link[base]{saveRDS}}.}

\item{method}{takes "network", "hclust" or "both" as a method}

\item{longMessages}{by default \code{FALSE} to keep the output short.}
}
\value{
placeholder
}
\description{
The essential wrapper function that runs all steps of the word
    analysis.
}
\examples{
\dontrun{
placeholder}
}
\seealso{
Other ginko functions: 
\code{\link{calculateModels}()},
\code{\link{createOrdinationPlot}()},
\code{\link{createTextMatrixFromPDF}()},
\code{\link{inspectGinko}()},
\code{\link{mostImportantPaperPerCluster}()},
\code{\link{processMetaDataMatrix}()}
}
\author{
Matthias Nachtmann, \email{matthias.nachtmann@stud.leuphana.de},
    Lisa Gotzian, \email{lisa.gotzian@stud.leuphana.de}
}
\concept{ginko functions}
