% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_processMetaDataMatrix.R
\name{processMetaDataMatrix}
\alias{processMetaDataMatrix}
\title{processMetaDataMatrix}
\usage{
processMetaDataMatrix(
  metaMatrix,
  control = list(),
  ignoreWords = c(),
  keepWordsFile
)
}
\arguments{
\item{metaMatrix}{metaMatrix created through \code{\link[scicloud]{getScopusMetaData}}
or \code{\link[scicloud]{createTextMatrixFromPDF}}. Equations, symbols, all words
in parentheses and all references are removed.}

\item{control}{a list of parameters used to determine preprocessing methods.
Error will be thrown if language & stemWords are not defined.
language: this defines the stopwords to be filtered. the default is
"english". Look at \code{\link[tm]{stopwords}} for more information.
stemWords: can be \code{TRUE} of \code{FALSE}. Transforms every word
to its stem, so variants of the same words are treated equally. Look
at \code{\link[tm]{stemDocument}} for more information.
saveToWd: a logical parameter whether or not to save the output of the function to the working directory
ordinationFunction placeholder
e.g. control = list(language = "SMART", stemWords = FALSE, saveToWd = TRUE, ordinationFunction = FALSE)}

\item{ignoreWords}{a vector of words to be ignored.}

\item{keepWordsFile}{path to a .csv-file that specifies which words to keep
during the analysis. Accepts 0/1 behind each word or takes the words
as they are and disregards all other words of the analysis. A template
for this can be generated with \code{generateWordlist} in
\code{\link[scicloud]{ordinationCluster}} or \code{\link[scicloud]{calculateModels}}.}
}
\value{
returns a list object with \code{[["Tf_idf"]]} as the tf-idf document
    term matrix, \code{[["MetaMatrix"]]} as passed to the function and
\code{[["wordList"]]} is the list of all words found in the papers.
}
\description{
The second function to the word analysis with scicloud This function
    accepts a \code{metaMatrix} and constructs a tf-idf matrix.
}
\seealso{
\code{\link[scicloud]{createTextMatrixFromPDF}} for the preceding step,
    \code{\link[scicloud]{calculateModels}} for the proceeding step

Other scicloud functions: 
\code{\link{calculateModels}()},
\code{\link{createOrdinationPlot}()},
\code{\link{createTextMatrixFromPDF}()},
\code{\link{delete_RDS}()},
\code{\link{getScopusMetaData}()},
\code{\link{inspectScicloud}()},
\code{\link{mostImportantPaperPerCluster}()},
\code{\link{ordinationCluster}()},
\code{\link{searchScopus}()}
}
\author{
Jia Yan Ng, \email{jia.y.ng@stud.leuphana.de}
}
\concept{scicloud functions}
